Pong UI'ing for some reason

+++ srcs/nodejs_base_image/utils/api/service/common/zodRules.ts
@@ -1,7 +1,6 @@
 import { z } from "zod";
 
 // Ids are always positive integers
-export const idValue = z.coerce.number().min(1);
-export const gameIdValue = z.coerce.number().min(1);
-export const userIdValue = z.coerce.number().min(1);
-
+export const idValue = z.coerce.number().int().min(1);
+export const gameIdValue = z.coerce.number().int().min(1);
+export const userIdValue = z.coerce.number().int().min(1);
